Makefile: Build with LuaJIT instead of minilua
RaptorJIT now uses LuaJIT for bootstrapping i.e. for code generation
during the build process (DynASM etc.)

This means that LuaJIT needs to be available (in PATH) at build time.
It also means that minilua.c is removed from the source tree.

The next step will be to formalize a build-time dependency on LuaJIT
2.0.4 in a nix expression (see RaptorJIT/RaptorJIT#25).

(Considered using PUC Lua for bootstrapping but then we need to also
install the BitOps library for compatibility with DynASM and I see no
immediate practical advantage over LuaJIT.)