Remove LJ_LIKELY() and LJ_UNLIKELY()
I find that these declarations clutter the code, and I consider them
to be optimizations that should only be used when and where they have
a beneficial effect. I also suspect the benefit to be rare on
supported RaptorJIT platforms because:

- Intel CPUs do excellent dynamic branch prediction;
- Declarations don't affect the interpreter (asm code);
- Declarations don't affect the JIT (generated code)

So it's only the runtime system e.g. garbage collector.

Have to check for a performance regression before merging this change.

(I am looking for the LKML thread where Ingo Molnar benchmarked the
kernel and found that the likely/unlikely annotations were actually
_reducing_ performance in practice -- but I can't find it so maybe I
dreamed it.)