lj_opt_sink.c: Eliminate GCSTEP when made redundant by sinking
Side traces that start by importing GC objects from their parents
include a GCSTEP instruction immediately after snapshot 0.

If the allocations of all of these objects are sunk then the GCSTEP
becomes redundant.

This change eliminates such redundant GCSTEP instructions by
converting them to NOP.

Note: This only affects the IR and not the generated machine code. The
assembler includes an equivalent optimization in lj_asm.c already.
Therefore the value of this change is not to increase performance but
only to simplify IR listings and make them correspond more closely
with generated code i.e. only show the GCSTEP IR instruction when a
call to lj_gc_step_jit() is actually going to be emitted.