Added PROFESSIONAL-SUPPORT.md list of service providers
RaptorJIT users need to know that there is professional support
available if and when they need it. This is to reduce the risks
associated with starting new projects (e.g. missing deadlines due to
learning curve) and to reduce the costs of maintaining these
applications (e.g. having in-house JIT internals gurus who need to be
hired/trained/retained/replaced/etc over time.)