Allow sinking CNEWI instructions that escape the last snapshot
Generally allocations that are referenced in the last snapshot of a
non-looping trace, i.e. values that will be committed to the Lua
stack on successful completion of the trace, are not allowed to sink.

This change introduces an exception so that CNEWI instructions are
allowed to sink when the linktype is to a root trace (but not to the
interpreter.) This is a preliminary change for future work towards
allowing unsunk values to move between root traces. (For now it just
causes Lua stack corruption.)