lj_vm.c: fix FORI/FORL bytecodes to support basic iteration
I do not actually understand how the original x86 ASM code for these bytecodes
is supposed to work. In light of that, and in order to make quick progress, I
changed FORI/FORL to implement the behavior I think they should have. I also
decided not to follow the fall-through style used in the ASM x86 VM and the
prior, broken implementation for reasons of clarity.