lj_trace.c: clear all hotcounts every second
We reset all hotcounts every second. This is a rough way to establish a
relation with elapsed time so that hotcounts provide a measure of frequency.

The concrete goal is to ensure that the JIT will trace code that becomes hot
over a short duration, but not code that becomes hot over, say, the course
of an hour.